requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.121.2
h11==0.16.0
idna==3.11
mysql-connector-python==9.5.0
passlib==1.7.4
pydantic==2.12.4
pydantic_core==2.41.5
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
